The `moe_align_block_size` is used by many moe models (e.g. DeepSeek-V3) but it is not compliable with cuda graph now, and its efficiency is relatively low when `num_experts` is high.

This PR optimizes `moe_align_block_size` from three aspects:

1. Replace `cudaMalloc` with `torch::empty`, so that we can enable cuda graph for global memory kernel.
2. When the num_tokens is small (in generation stage, this is the common case), the nelements of `topk_ids` should always smaller than 65535 (max value of `uint16`), so we can use `uint16` as the  dtypes of `tokens_cnts`, this requires only half of the shared memory size. With this change, DeepSeek-V3 can use shared memory kernel instead of global memory kernel.
3. Remove hard-code `num_experts >= 256`, we can choose the best kernel based on the input and the maximum shared memory size of the current device.


| setting | generation speed |
| - | - |
| awq + marlin-moe-kernel + no-cuda-graph | 5.4tok/s |
| awq + marlin-moe-kernel + cuda-graph | 10.5tok/s |
| awq + marlin-moe-kernel + cuda-graph + shared_memory_kernel | 11.1tok/s |
| awq + triton-fused-moe-kernel + no-cuda-graph | 10.0tok/s |
| awq + triton-fused-moe-kernel + cuda-graph | 25.3tok/s |
| awq + triton-fused-moe-kernel + cuda-graph + shared_memory_kernel | 29.6tok/s |


( triton-fused-moe-kernel is from https://github.com/vllm-project/vllm/pull/12185 )