@mgoin 2025-01-21
Verified with gsm8k eval

```
lm_eval --model vllm --model_args pretrained=/data/xmo/whales,trust_remote_code=True,tensor_parallel_size=8,max_model_len=4096 --tasks gsm8k --num_fewshot 5 --batch_size auto

Processed prompts: 100%|███████████████████████| 1319/1319 [06:17<00:00,  3.49it/s, est. speed input: 3048.29 toks/s, output: 330.31 toks/s]
Running generate_until requests: 100%|██████████████████████| 1319/1319 [06:17<00:00,  3.49it/s]

vllm (pretrained=/data/xmo/whales,trust_remote_code=True,tensor_parallel_size=8,max_model_len=4096), gen_kwargs: (None), limit: None, num_fewshot: 5, batch_size: auto
|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.9295|±  |0.0071|
|     |       |strict-match    |     5|exact_match|↑  |0.9295|±  |0.0071|
```